Fixed bug in C solution of Factorial-20.

+++ Factorial-20/Factorial20.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-long factorial( int value ) {
-  long result = 1;
+long long factorial( int value ) {
+  long long result = 1;
 
   for ( int iteration = 1; iteration <= value; iteration++ ) {
     result = result * iteration;
@@ -13,8 +13,8 @@ long factorial( int value ) {
 int main( int argc, char **argv ) {
   for ( int i = 0; i <= 50000; i++ ) {
     for ( int value = 0; value <= 20; value++ ) {
-      long result = factorial( value );
-      printf( "%ld\n", result );
+      long long result = factorial( value );
+      printf( "%lld\n", result );
     }
   }
 }
